practice/chapter14/code/requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
umap-learn>=0.5.4
gradio>=4.0.0
evaluate>=0.4.0
